Implemented a rough version of the multiplayer game
1)Implemented the system of progress for users to see.
2)Implemented multiplayer for up to 3 users (as far as I tested). Now users can play multiplayer against each other!!!
3)Made so that users cannot copy the text of the page (to cheat and finish faster)

I used the following code to achieve that:
	-webkit-touch-callout: none;
	-webkit-user-select: none;
	-khtml-user-select: none;
	-moz-user-select: none;
	-ms-user-select: none;
	user-select: none;

I spend the whole day having a lot of problems. One I was having was about the server storing two version of the quotes and authors. I solved that by making sure only one user triggers one event that takes the data to the server. Otherwise it won't work.